Moved tree declarations around, *mut -> *const
Moved the tree definition to the top, and the actual tree in the
lazy_static! lower down. Changed the *mut Node back to a *const Node, if
a mutable node is that should be done within the Node struct, not at the
tree level.

Commented out the main body of move_container_to_workspace, it requires
the aforementioned mutability of the active container to work.