Fixed workspace-switching panic, with a hack
When switching workspaces and deleting the previous one, it keeps the
previous one in a bad state because it assumes the active container is
part of the workspace being deleted since we are in the middle of
switching workspaces. We set the active workspace to none, and that
fixes that part of the problem.

The other part of the problem (which is the hack) involves the
consistency of the tree at the end of the remove_view_or_container
function. When you remove a workspace, the tree exists in a non-valid
state. Because it deletes the nodes newest to oldest, the workspace is
usually itself is usually the last thing to go, so it will be in the
tree and have no children, which is against the notion of the tree.

Another function will have to be made to remedy this, to explicitly not
do a check and only work on workspaces.