Lua thread: Run a GLib GMainLoop
This makes e.g. the following Lua code work:

    require("gears.timer").start_new(1, function()
        print("I live!")
        return true
    end)

Instead of using an mpsc channel for communicating with the Lua thread,
this now uses glib::source::idle_add() to a run a closure inside the
GLib main loop, which in turn runs inside of the Lua thread. Thus, the
closure can then access the LUA global to actually run Lua code.

Signed-off-by: Uli Schlachter <psychon@znc.in>