Emit refresh signal in Lua thread
In awesome, at the end of each main loop iteration the "refresh" signal
is emitted. This is used by gears.timer.delayed_call() to do, well,
delayed calls. This is in turn used to redraw wiboxes with the
intentions that the redraw only happens after all changes to the widgets
were done, not for each single change.

To make this work in way-cooler, the signal is emitted whenever some Lua
code was run.

Signed-off-by: Uli Schlachter <psychon@znc.in>